Get rid of AbstractProcess
It now only had one subclass. I'd have got rid of the subclass and used
the name of the superclass instead, except that the subclass has
references to it in SpiNNFrontEndCommon
